+++ src/App.vue
@@ -10,7 +10,7 @@ import TransactionModal from "@/components/TransactionModal.vue";
 // Views
 import DashboardView from "@/views/DashboardView.vue";
 import TransactionsView from "@/views/TransactionsView.vue";
-import CategoriesView from "@/views/CategoriesView.vue";
+import CategoriesView from "@/views/labels/LabelsView.vue";
 import SettingsView from "@/views/settings/SettingsView.vue";
 import AccountsView from "@/views/accounts/AccountsView.vue";
 import InsightsView from "@/views/insights/InsightsView.vue";
